ritza-real-world-tests/browserbase/requirements.txt
playwright
python-dotenv
browserbase
